feat: improves empty state UI for api-keys page and refactors it into reusable section component
Improves maintainability by moving API keys UI logic and table features into a dedicated section component. Enhances empty state with better visuals and links to documentation, and streamlines main page to use new abstraction.
